# contracts/operations/v1/openapi.yaml
openapi: 3.1.0
info:
  title: BoatOps Internal Operations API
  version: 1.0.0-alpha.5
  summary: Private operations commands and operator schedule catalog/calendar reads.
  description: >-
    Private alpha contract. Inventory and trip commands require operations.write.
    Finance writes require operations.finance.write, and finance reads require
    operations.finance.read or operations.finance.write. Schedule reads require
    operations.schedule.read or operations.schedule.write; schedule writes require
    operations.schedule.write. Schedule catalog writes use canonical server-owned
    identities and audit logs. This API is not part of the channel-facing Inventory
    Provider API and must not be exposed to untrusted channel clients.
servers:
  - url: https://boatops.ayany.com
    description: Ayany private alpha target
  - url: http://localhost:8000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/internal/v1/fuel-logs:
    post:
      tags: [Operating costs]
      operationId: recordFuelLog
      summary: Record a posted fuel purchase for a boat and optional trip
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RecordFuelLogRequest'}
      responses:
        '201':
          description: Fuel log posted and audited.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FuelLogResponse'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/internal/v1/expenses:
    post:
      tags: [Operating costs]
      operationId: recordExpense
      summary: Record one posted operating expense with one or more category lines
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RecordExpenseRequest'}
      responses:
        '201':
          description: Expense and lines posted atomically.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExpenseResponse'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/internal/v1/stock/movements:
    post:
      tags: [Operating costs]
      operationId: recordStockMovement
      summary: Post a purchase, load, consumption, return, waste, or adjustment movement
      description: >-
        The item row is locked while balances are updated. LOAD and RETURN transfer cost
        between warehouse and boat balances. CONSUME and WASTE recognize cost at the
        source moving-average unit cost. A movement that would create negative stock fails.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RecordStockMovementRequest'}
      responses:
        '201':
          description: Stock movement and all affected balances posted atomically.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StockMovementResponse'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/internal/v1/fuel-logs/{id}:reverse:
    post:
      tags: [Operating costs]
      operationId: reverseFuelLog
      summary: Reverse one posted fuel log
      parameters:
        - $ref: '#/components/parameters/ObjectId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReverseRecordRequest'}
      responses:
        '201': {$ref: '#/components/responses/ReversalCreated'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/internal/v1/expenses/{id}:reverse:
    post:
      tags: [Operating costs]
      operationId: reverseExpense
      summary: Reverse one posted expense
      parameters:
        - $ref: '#/components/parameters/ObjectId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReverseRecordRequest'}
      responses:
        '201': {$ref: '#/components/responses/ReversalCreated'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/internal/v1/stock/movements/{id}:reverse:
    post:
      tags: [Operating costs]
      operationId: reverseStockMovement
      summary: Reverse one posted stock movement
      description: Posts one exact-value REVERSAL compensation atomically. Compensations cannot be reversed.
      parameters:
        - $ref: '#/components/parameters/ObjectId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReverseRecordRequest'}
      responses:
        '201': {$ref: '#/components/responses/ReversalCreated'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque API token with explicit operations, finance, or schedule scope
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
        minLength: 36
        maxLength: 255
    ObjectId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    AuthorizationFailed:
      description: Authentication, organization isolation, or required operation scope failed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Conflict:
      description: Inventory, state, readiness, external-reference, or idempotency conflict.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    ValidationFailed:
      description: Request validation failed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    ReversalCreated:
      description: Reversal, audit, idempotency record, and status or compensation committed atomically.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/FinanceReversalResponse'}
  schemas:
...
    RecordFuelLogRequest:
      type: object
      additionalProperties: false
      required: [external_reference, boat_id, cash_account_id, occurred_at, station_name, liters, price_per_liter_minor, total_amount_minor, currency, handled_by]
      properties:
        external_reference: {type: string, minLength: 1, maxLength: 255}
        boat_id: {type: integer, minimum: 1}
        trip_id: {type: [integer, 'null'], minimum: 1}
        cash_account_id: {type: integer, minimum: 1}
        occurred_at: {type: string, format: date-time}
        station_name: {type: string, minLength: 1, maxLength: 255}
        liters: {type: number, exclusiveMinimum: 0, multipleOf: 0.001}
        price_per_liter_minor: {type: integer, minimum: 0}
        total_amount_minor:
          type: integer
          minimum: 0
          description: Must equal liters multiplied by price_per_liter_minor within one minor unit of rounding.
        currency: {type: string, pattern: '^[A-Z]{3}$'}
        fuel_level_before_percent: {type: [number, 'null'], minimum: 0, maximum: 100}
        fuel_level_after_percent: {type: [number, 'null'], minimum: 0, maximum: 100}
        engine_hours: {type: [number, 'null'], minimum: 0}
        handled_by: {type: string, minLength: 1, maxLength: 255}
        receipt_reference: {type: [string, 'null'], maxLength: 500}
        notes: {type: [string, 'null'], maxLength: 2000}
    ExpenseLineInput:
      type: object
      additionalProperties: false
      required: [expense_category_id, description, amount_minor]
      properties:
        expense_category_id: {type: integer, minimum: 1}
        description: {type: string, minLength: 1, maxLength: 500}
        amount_minor: {type: integer, minimum: 0}
    RecordExpenseRequest:
      type: object
      additionalProperties: false
      required: [external_reference, cash_account_id, occurred_at, currency, handled_by, lines]
      properties:
        external_reference: {type: string, minLength: 1, maxLength: 255}
        boat_id:
          type: [integer, 'null']
          minimum: 1
          description: Required when any line is DIRECT or BOAT_DAILY; inferred from trip_id when omitted.
        trip_id: {type: [integer, 'null'], minimum: 1}
        cash_account_id: {type: integer, minimum: 1}
        occurred_at: {type: string, format: date-time}
        currency: {type: string, pattern: '^[A-Z]{3}$'}
        handled_by: {type: string, minLength: 1, maxLength: 255}
        receipt_reference: {type: [string, 'null'], maxLength: 500}
        notes: {type: [string, 'null'], maxLength: 2000}
        lines:
          type: array
          minItems: 1
          maxItems: 50
          items: {$ref: '#/components/schemas/ExpenseLineInput'}
    RecordStockMovementRequest:
      type: object
      additionalProperties: false
      required: [external_reference, item_id, movement_type, occurred_at, quantity, handled_by]
      properties:
        external_reference: {type: string, minLength: 1, maxLength: 255}
        item_id: {type: integer, minimum: 1}
        boat_id:
          type: [integer, 'null']
          minimum: 1
          description: Required for LOAD, CONSUME, and RETURN; optional stock location for waste and adjustments.
        trip_id:
          type: [integer, 'null']
          minimum: 1
          description: Allowed for LOAD, CONSUME, RETURN, and WASTE; the trip determines the boat when boat_id is omitted.
        cash_account_id:
          type: [integer, 'null']
          minimum: 1
          description: Required only for PURCHASE and must use the item's currency.
        movement_type:
          type: string
          enum: [PURCHASE, LOAD, CONSUME, RETURN, WASTE, ADJUSTMENT_IN, ADJUSTMENT_OUT]
        occurred_at: {type: string, format: date-time}
        quantity: {type: number, exclusiveMinimum: 0, multipleOf: 0.001}
        total_cost_amount_minor:
          type: [integer, 'null']
          minimum: 0
          description: Required only for PURCHASE and ADJUSTMENT_IN; other movement costs come from moving-average cost.
        handled_by: {type: string, minLength: 1, maxLength: 255}
        receipt_reference: {type: [string, 'null'], maxLength: 500}
        reason:
          type: [string, 'null']
          maxLength: 2000
          description: Required for WASTE, ADJUSTMENT_IN, and ADJUSTMENT_OUT.
    ReverseRecordRequest:
      type: object
      additionalProperties: false
      required: [external_reference, reason]
      properties:
        external_reference: {type: string, minLength: 1, maxLength: 255}
        reason: {type: string, minLength: 3, maxLength: 2000}
    FinanceReversalResponse:
      type: object
      additionalProperties: true
      required: [request_id, idempotency_key, organization_id, finance_reversal_id, original_record_type, original_record_id, status, code, occurred_at, business_timezone]
      properties:
        request_id: {type: string, format: uuid}
        idempotency_key: {type: string}
        organization_id: {type: integer, minimum: 1}
        finance_reversal_id: {type: integer, minimum: 1}
        original_record_type: {type: string, enum: [fuel_log, expense, stock_movement]}
        original_record_id: {type: integer, minimum: 1}
        compensation: {type: [object, "null"]}
        status: {const: REVERSED}
        code: {const: FINANCE_RECORD_REVERSED}
        occurred_at: {type: string, format: date-time}
        business_timezone: {type: string}
...
    FuelLogResponse:
      type: object
      additionalProperties: false
      required: [request_id, idempotency_key, organization_id, fuel_log_id, external_reference, boat_id, trip_id, liters, total_amount_minor, currency, status, code, occurred_at, business_timezone]
      properties:
        request_id: {type: string, format: uuid}
        idempotency_key: {type: string}
        organization_id: {type: integer, minimum: 1}
        fuel_log_id: {type: integer, minimum: 1}
        external_reference: {type: string}
        boat_id: {type: integer, minimum: 1}
        trip_id: {type: [integer, 'null'], minimum: 1}
        liters: {type: string, pattern: '^\d+\.\d{3}$'}
        total_amount_minor: {type: integer, minimum: 0}
        currency: {type: string, pattern: '^[A-Z]{3}$'}
        status: {const: POSTED}
        code: {const: FUEL_LOG_RECORDED}
        occurred_at: {type: string, format: date-time}
        business_timezone: {type: string}
    ExpenseResponse:
      type: object
      additionalProperties: false
      required: [request_id, idempotency_key, organization_id, expense_id, external_reference, boat_id, trip_id, line_count, total_amount_minor, currency, status, code, occurred_at, business_timezone]
      properties:
        request_id: {type: string, format: uuid}
        idempotency_key: {type: string}
        organization_id: {type: integer, minimum: 1}
        expense_id: {type: integer, minimum: 1}
        external_reference: {type: string}
        boat_id: {type: [integer, 'null'], minimum: 1}
        trip_id: {type: [integer, 'null'], minimum: 1}
        line_count: {type: integer, minimum: 1}
        total_amount_minor: {type: integer, minimum: 0}
        currency: {type: string, pattern: '^[A-Z]{3}$'}
        status: {const: POSTED}
        code: {const: EXPENSE_RECORDED}
        occurred_at: {type: string, format: date-time}
        business_timezone: {type: string}
    StockMovementResponse:
      type: object
      additionalProperties: false
      required: [request_id, idempotency_key, organization_id, movement_id, external_reference, item_id, boat_id, trip_id, cash_account_id, movement_type, quantity, unit_cost_minor, total_cost_amount_minor, currency, from_location_key, to_location_key, from_balance_quantity, to_balance_quantity, status, code, occurred_at, business_timezone]
      properties:
        request_id: {type: string, format: uuid}
        idempotency_key: {type: string}
        organization_id: {type: integer, minimum: 1}
        movement_id: {type: integer, minimum: 1}
        external_reference: {type: string}
        item_id: {type: integer, minimum: 1}
        boat_id: {type: [integer, 'null'], minimum: 1}
        trip_id: {type: [integer, 'null'], minimum: 1}
        cash_account_id: {type: [integer, 'null'], minimum: 1}
        movement_type:
          type: string
          enum: [PURCHASE, LOAD, CONSUME, RETURN, WASTE, ADJUSTMENT_IN, ADJUSTMENT_OUT]
        quantity: {type: string, pattern: '^\d+\.\d{3}$'}
        unit_cost_minor: {type: string, pattern: '^\d+\.\d{6}$'}
        total_cost_amount_minor: {type: integer, minimum: 0}
        currency: {type: string, pattern: '^[A-Z]{3}$'}
        from_location_key: {type: [string, 'null']}
        to_location_key: {type: [string, 'null']}
        from_balance_quantity: {type: [string, 'null']}
        to_balance_quantity: {type: [string, 'null']}
        status: {const: POSTED}
        code: {const: STOCK_MOVEMENT_RECORDED}
        occurred_at: {type: string, format: date-time}
        business_timezone: {type: string}
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [request_id, code, message, retryable, manual_action_required]
      properties:
        request_id: {type: string, format: uuid}
        code: {type: string}
        message: {type: string}
        retryable: {type: boolean}
        manual_action_required: {type: boolean}